scd-proiect/worker-service/requirements.txt
pika
fpdf2
psycopg2-binary
